test(desktop): consolidate embedding availability coverage
